add custom text document content provider
