+++ package.json
@@ -182,12 +182,12 @@
     "prettier": "^2.0.5",
     "rimraf": "^3.0.2",
     "ts-node": "^3.3.0",
-    "typescript": "3.5.1",
+    "typescript": "4.1.2",
     "vscode-test": "^1.4.0"
   },
   "dependencies": {
     "request-light": "^0.4.0",
-    "vscode-languageclient": "5.2.1",
+    "vscode-languageclient": "7.0.0",
     "vscode-nls": "^3.2.1",
     "vscode-uri": "^2.0.3",
     "yaml-language-server": "next"
